worked on dg
