+++ docs/diagrams/tasks/addNewTask.puml
@@ -1,105 +1,42 @@
 @startuml
 
-activate ":Duke"
-":Duke" -> ":TaskManager"++ : execute
 
-loop true
+-> ":AddTask" : execute()
+activate ":AddTask"
+ref over ":AddTask" : get module
 
-":TaskManager" -> ":Ui"++ : printTaskManagerMenu
+":AddTask" -> ":Ui"++ : printAddTaskDescriptionMessage()
 return
-":TaskManager" -> ":Ui"++ : readCommand
+":AddTask" -> ":AddTask"++ : getDescription()
+return description
+
+opt taskNumber != 1
+":AddTask" -> ":AddTask"++ : getDateAndTime(taskTypeNumber)
+return dateAndTime
+end
+
+":AddTask" -> ":Ui"++ : printAddMessageAfterCompletedTask()
 return
+":AddTask" -> ":AddTask"++ : getMessage()
+return message
 
-alt Integer.parseInt(command)
-    opt command.equals(EXIT_COMMAND)
-    ":TaskManager" --> ":Duke"
-    end
-
-":TaskManager" -> ":TaskManager"++ : addNewTask
-":TaskManager" -> ":Ui"++ : printAddTaskMenu
+alt ADD_TASK_COMMAND
+":AddTask" -> ":AddTask"++ : addTask(module, description, message)
 return
-":TaskManager" -> ":TaskManager"++ : getTaskNumber
+else ADD_ASSIGNMENT_COMMAND
+":AddTask" -> ":AddTask"++ : addAssignment(module, description, message, dateAndTime)
 return
-":TaskManager" -> ":AddTask"++ : execute
-
-    alt ModuleInfo.modules.isEmpty()
-    ":AddTask" -> ":Ui"++ : printNoModuleMessage
-    return
-    ":AddTask" -> ":Ui"++ : readCommand
-    return
-        opt input.equalsIgnoreCase("N")
-        ":AddTask" --> ":TaskManager"
-        end
-    ":AddTask" -> ":ModuleInfo"++ : addNewModule
-    return
-    ":AddTask" -> ":ModuleInfo"++ : modules
-    ":ModuleInfo" -> ":Module"++ : getName
-    return
-    destroy ":Module"
-    return
-    destroy ":ModuleInfo"
-
-    else !ModuleInfo.modules.isEmpty()
-    ":AddTask" -> ":Ui"++ : printAddTaskModuleMessage
-    return
-    ":AddTask" -> ":AddTask"++ : printAndGetModule
-    return
-        opt module.equals("")
-        ":AddTask" --> ":TaskManager"
-        end
-    end
-
-    ":AddTask" -> ":Ui"++ : printAddTaskDescriptionMessage
-    return
-    ":AddTask" -> ":Ui"++ : readCommand
-    return
-
-    opt taskNumber != 1
-    ":AddTask" -> ":AddTask"++ : getDate
-    return
-    ":AddTask" -> ":AddTask"++ : getTime
-    return
-    end
-
-    ":AddTask" -> ":Ui"++ : printAddMessageAfterCompletedTask
-    return
-    ":AddTask" -> ":Ui"++ : readCommand
-    return
-
-    alt ADD_TASK_COMMAND
-    ":AddTask" -> ":AddTask"++ : addTask
-    return
-    else ADD_ASSIGNMENT_COMMAND
-    ":AddTask" -> ":AddTask"++ : addAssignment
-    return
-    else ADD_MIDTERM_COMMAND
-    ":AddTask" -> ":AddTask"++ : addMidterm
-    return
-    else ADD_FINAL_EXAM_COMMAND
-    ":AddTask" -> ":AddTask"++ : addFinalExam
-    return
-    else
-    ":AddTask" -> ":Ui"++ : printInvalidIntegerMessage
-    return
-    end
-
-    return
-    return
-
-else NumberFormatException
-":TaskManager" -> ":Ui"++ : printInvalidIntegerMessage
+else ADD_MIDTERM_COMMAND
+":AddTask" -> ":AddTask"++ : addMidterm(module, description, message, dateAndTime)
+return
+else ADD_FINAL_EXAM_COMMAND
+":AddTask" -> ":AddTask"++ : addFinalExam(module, description, message, dateAndTime)
+return
+else default
+":AddTask" -> ":Ui"++ : printRepeatInputUntilValidMessage()
 return
 end
 
-alt
-":TaskManager" -> ":Storage"++ : saveAllFiles
-return
-else IOException
-":TaskManager" -> ":Ui"++ : printFilesCouldNotBeSavedMessage
-return
-end
-
-end
 return
 
 @enduml
