+++ docs/diagrams/tasks/markUnmarkTask.puml
@@ -1,71 +1,42 @@
 @startuml
 
-activate ":Duke"
-":Duke" -> ":TaskManager"++ : execute
-
-loop true
-
-":TaskManager" -> ":Ui"++ : printTaskManagerMenu
-return
-":TaskManager" -> ":Ui"++ : readCommand
-return
-
-alt Integer.parseInt(command)
-    opt command.equals(EXIT_COMMAND)
-    ":TaskManager" --> ":Duke"
-    end
-
-":TaskManager" -> ":TaskManager"++ : markOrUnmarkTask
-":TaskManager" -> ":Ui"++ : printMarkTaskMenu
-return
-":TaskManager" -> ":TaskManager"++ : getTaskNumber
-return
-":TaskManager" -> ":MarkOrUnmarkTask"++ : execute
+-> ":MarkOrUnmarkTask" : execute()
+activate ":MarkOrUnmarkTask"
 
 opt TaskManager.taskListIsEmpty(taskTypeNumber)
-    ":MarkOrUnmarkTask" -> ":Ui"++ : printTaskListIsEmptyMessage
-    return
-    ":MarkOrUnmarkTask" --> ":TaskManager"
-    end
-    ":MarkOrUnmarkTask" -> ":Ui"++ : printSelectTaskNumberToMarkOrUnmark
-    return
-    alt TOGGLE_TASK_COMMAND
-    ":MarkOrUnmarkTask" -> ":MarkOrUnmarkTask"++ : toggleTaskStatus
-    return
-    else TOGGLE_ASSIGNMENT_COMMAND
-    ":MarkOrUnmarkTask" -> ":MarkOrUnmarkTask"++ : toggleTaskStatus
-    return
-    else TOGGLE_MIDTERM_COMMAND
-    ":MarkOrUnmarkTask" -> ":MarkOrUnmarkTask"++ : toggleTaskStatus
-    return
-    else TOGGLE_FINAL_EXAM_COMMAND
-    ":MarkOrUnmarkTask" -> ":MarkOrUnmarkTask"++ : toggleTaskStatus
-    return
-    else default
-    ":MarkOrUnmarkTask" -> ":Ui"++ : printInvalidIntegerMessage
-    return
-    end
+":MarkOrUnmarkTask" -> ":Ui"++ : printTaskListIsEmptyMessage
+return
+<---- ":MarkOrUnmarkTask"
+end
 
-else NumberFormatException
-":MarkOrUnmarkTask" -> ":Ui"++ : printInvalidIntegerMessage
+":MarkOrUnmarkTask" -> ":Ui"++ : printSelectTaskNumberToMarkOrUnmark(taskTypeNumber)
 return
 
-else IndexOutOfBoundsException
-":MarkOrUnmarkTask" -> ":Ui"++ : printInvalidTaskNumberMessage
+loop
+":MarkOrUnmarkTask" -> ":Ui"++ : readCommandToInt()
+return taskNumber
+
+alt TOGGLE_TASK_COMMAND
+":MarkOrUnmarkTask" -> ":MarkOrUnmarkTask"++ : toggleTaskStatus(taskNumber, TASK_TYPE)
+return
+<---- ":MarkOrUnmarkTask"
+else TOGGLE_ASSIGNMENT_COMMAND
+":MarkOrUnmarkTask" -> ":MarkOrUnmarkTask"++ : toggleTaskStatus(taskNumber, ASSIGNMENT_TYPE)
+return
+<--- ":MarkOrUnmarkTask"
+else TOGGLE_MIDTERM_COMMAND
+":MarkOrUnmarkTask" -> ":MarkOrUnmarkTask"++ : toggleTaskStatus(taskNumber, MIDTERM_TYPE)
+return
+<--- ":MarkOrUnmarkTask"
+else TOGGLE_FINAL_EXAM_COMMAND
+":MarkOrUnmarkTask" -> ":MarkOrUnmarkTask"++ : toggleTaskStatus(taskNumber, FINAL_EXAM_TYPE)
+return
+<--- ":MarkOrUnmarkTask"
+else default
+":MarkOrUnmarkTask" -> ":Ui"++ : printRepeatInputUntilValidMessage()
 return
 end
 
-return
-return
-
-alt
-":TaskManager" -> ":Storage"++ : saveAllFiles
-return
-else IOException
-":TaskManager" -> ":Ui"++ : printFilesCouldNotBeSavedMessage
-return
 end
-
-return
 
 @enduml
