+++ docs/diagrams/tasks/pinTask.puml
@@ -1,73 +1,41 @@
 @startuml
 
-activate ":Duke"
-":Duke" -> ":TaskManager"++ : execute
+-> ":PinTask"
+activate ":PinTask"
 
-loop true
-
-":TaskManager" -> ":Ui"++ : printTaskManagerMenu
+opt TaskManager.taskListIsEmpty(taskTypeNumber)
+":PinTask" -> ":Ui"++ : printTaskListIsEmptyMessage()
 return
-":TaskManager" -> ":Ui"++ : readCommand
-return
+<---- ":PinTask"
+end
 
-alt Integer.parseInt(command)
-    opt command.equals(EXIT_COMMAND)
-    ":TaskManager" --> ":Duke"
-    end
-
-":TaskManager" -> ":TaskManager"++ : pinTask
-":TaskManager" -> ":Ui"++ : rintPinTaskMenu
-return
-":TaskManager" -> ":TaskManager"++ : getTaskNumber
-return
-":TaskManager" -> ":PinTask"++ : execute
-    opt TaskManager.taskListIsEmpty(taskTypeNumber)
-    ":PinTask" -> ":Ui"++ : printTaskListIsEmptyMessage
-    return
-    ":PinTask" --> ":TaskManager"
-    end
-    ":PinTask" -> ":Ui"++ : printSelectTaskNumberToPin
-    return
-    alt PIN_TASK_COMMAND
-    ":PinTask" -> ":PinTask"++ : addTaskToPinnedTasks
-    return
-    else PIN_ASSIGNMENT_COMMAND
-    ":PinTask" -> ":PinTask"++ : addTaskToPinnedTasks
-    return
-    else PIN_MIDTERM_COMMAND
-    ":PinTask" -> ":PinTask"++ : addTaskToPinnedTasks
-    return
-    else PIN_FINAL_EXAM_COMMAND
-    ":PinTask" -> ":PinTask"++ : addTaskToPinnedTasks
-    return
-    else default
-    ":PinTask" -> ":Ui"++ : printInvalidIntegerMessage
-    return
-    end
-
-else NumberFormatException
-":PinTask" -> ":Ui"++ : printInvalidIntegerMessage
+":PinTask" -> ":Ui"++ : printSelectTaskNumberToPin(taskTypeNumber)
 return
 
-else IndexOutOfBoundsException
-":PinTask" -> ":Ui"++ : printInvalidTaskNumberMessage
+loop
+":PinTask" -> ":Ui"++ : readCommandToInt()
+return taskNumber
+
+alt PIN_TASK_COMMAND
+":PinTask" -> ":PinTask"++ : addTaskToPinnedTasks(TaskManager.tasks.get(taskNumber - 1), TASK_TYPE)
+return
+<---- ":PinTask"
+else PIN_ASSIGNMENT_COMMAND
+":PinTask" -> ":PinTask"++ : addTaskToPinnedTasks(TaskManager.assignments.get(taskNumber - 1), ASSIGNMENT_TYPE)
+return
+<---- ":PinTask"
+else PIN_MIDTERM_COMMAND
+":PinTask" -> ":PinTask"++ : addTaskToPinnedTasks(TaskManager.midterms.get(taskNumber - 1), MIDTERM_TYPE)
+return
+<---- ":PinTask"
+else PIN_FINAL_EXAM_COMMAND
+":PinTask" -> ":PinTask"++ : addTaskToPinnedTasks(TaskManager.finalExams.get(taskNumber - 1), FINAL_EXAM_TYPE)
+return
+<---- ":PinTask"
+else default
+":PinTask" -> ":Ui"++ : printRepeatInputUntilValidMessage()
 return
 end
-destroy ":PinTask"
-
-return
-return
-
-alt
-":TaskManager" -> ":Storage"++ : saveAllFiles
-return
-else IOException
-":TaskManager" -> ":Ui"++ : printFilesCouldNotBeSavedMessage
-return
-end
-destroy ":TaskManager"
-
-return
 
 end
