+++ src/main/java/seedu/duke/features/moduleinfo/Module.java
@@ -1,6 +1,6 @@
 package seedu.duke.features.moduleinfo;
 
-import seedu.duke.features.task.Task;
+import seedu.duke.features.task.tasktypes.Task;
 
 import java.util.ArrayList;
 import java.util.Hashtable;
