+++ src/main/java/seedu/duke/features/moduleinfo/ModuleInfo.java
@@ -5,10 +5,10 @@ import seedu.duke.storage.Storage;
 import seedu.duke.ui.Ui;
 import seedu.duke.features.capsimulator.ModuleGradeEnum;
 import seedu.duke.features.link.Links;
-import seedu.duke.features.task.Assignment;
-import seedu.duke.features.task.FinalExam;
-import seedu.duke.features.task.Midterm;
-import seedu.duke.features.task.Task;
+import seedu.duke.features.task.tasktypes.Assignment;
+import seedu.duke.features.task.tasktypes.FinalExam;
+import seedu.duke.features.task.tasktypes.Midterm;
+import seedu.duke.features.task.tasktypes.Task;
 import seedu.duke.features.task.TaskManager;
 import seedu.duke.features.task.command.DeleteTask;
