+++ src/main/java/seedu/duke/features/task/tasktypes/Task.java
@@ -1,4 +1,4 @@
-package seedu.duke.features.task;
+package seedu.duke.features.task.tasktypes;
 
 //@@author hazelhedmine-reused
 //Reused from https://github.com/hazelhedmine/ip/blob/master/src/main/java/duke/task/Task.java
