+++ src/main/java/seedu/duke/storage/Storage.java
@@ -8,10 +8,10 @@ import seedu.duke.features.link.LinkLoadException;
 import seedu.duke.features.link.ZoomLinkInfo;
 import seedu.duke.features.moduleinfo.Module;
 import seedu.duke.features.moduleinfo.ModuleInfo;
-import seedu.duke.features.task.Assignment;
-import seedu.duke.features.task.FinalExam;
-import seedu.duke.features.task.Task;
-import seedu.duke.features.task.Midterm;
+import seedu.duke.features.task.tasktypes.Assignment;
+import seedu.duke.features.task.tasktypes.FinalExam;
+import seedu.duke.features.task.tasktypes.Task;
+import seedu.duke.features.task.tasktypes.Midterm;
 import seedu.duke.features.task.TaskManager;
 
 import java.io.File;
