+++ src/main/java/seedu/duke/ui/Ui.java
@@ -4,10 +4,10 @@ import seedu.duke.features.capsimulator.HelpGraduation;
 import seedu.duke.features.link.LinkInfo;
 import seedu.duke.features.link.ZoomLinkInfo;
 import seedu.duke.features.moduleinfo.ModuleInfo;
-import seedu.duke.features.task.Assignment;
-import seedu.duke.features.task.FinalExam;
-import seedu.duke.features.task.Midterm;
-import seedu.duke.features.task.Task;
+import seedu.duke.features.task.tasktypes.Assignment;
+import seedu.duke.features.task.tasktypes.FinalExam;
+import seedu.duke.features.task.tasktypes.Midterm;
+import seedu.duke.features.task.tasktypes.Task;
 import seedu.duke.features.task.TaskManager;
 
 import java.util.ArrayList;
